+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/controlplane/dispatcher/DefaultStreamManagerRunnerFactory.java
@@ -22,7 +22,7 @@ import org.apache.flink.configuration.Configuration;
 import org.apache.flink.runtime.dispatcher.DispatcherGateway;
 import org.apache.flink.runtime.webmonitor.retriever.LeaderGatewayRetriever;
 import org.apache.flink.streaming.controlplane.streammanager.StreamManagerConfiguration;
-import org.apache.flink.runtime.controlplane.streammanager.StreamManagerRunner;
+import org.apache.flink.streaming.controlplane.streammanager.StreamManagerRunner;
 import org.apache.flink.streaming.controlplane.streammanager.StreamManagerRunnerImpl;
 import org.apache.flink.streaming.controlplane.streammanager.factories.DefaultStreamManagerServiceFactory;
 import org.apache.flink.streaming.controlplane.streammanager.factories.StreamManagerServiceFactory;
