+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/controlplane/dispatcher/StreamManagerRunnerFactory.java
@@ -19,7 +19,7 @@
 package org.apache.flink.streaming.controlplane.dispatcher;
 
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.runtime.controlplane.streammanager.StreamManagerRunner;
+import org.apache.flink.streaming.controlplane.streammanager.StreamManagerRunner;
 import org.apache.flink.runtime.dispatcher.DispatcherGateway;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServices;
 import org.apache.flink.runtime.jobgraph.JobGraph;
