+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/controlplane/streammanager/StreamManager.java
@@ -22,26 +22,19 @@ import org.apache.flink.api.common.JobID;
 import org.apache.flink.api.common.time.Time;
 import org.apache.flink.runtime.clusterframework.types.ResourceID;
 import org.apache.flink.runtime.concurrent.FutureUtils;
-import org.apache.flink.runtime.controlplane.streammanager.StreamManagerAddress;
+import org.apache.flink.runtime.controlplane.streammanager.StreamManagerId;
 import org.apache.flink.runtime.dispatcher.DispatcherGateway;
 import org.apache.flink.runtime.controlplane.streammanager.StreamManagerGateway;
-import org.apache.flink.runtime.controlplane.streammanager.StreamManagerId;
-import org.apache.flink.runtime.controlplane.streammanager.StreamManagerService;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServices;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.runtime.jobmaster.JobMasterGateway;
 import org.apache.flink.runtime.jobmaster.JobMasterId;
 import org.apache.flink.runtime.jobmaster.JobMasterRegistrationSuccess;
-import org.apache.flink.runtime.leaderretrieval.LeaderRetrievalService;
 import org.apache.flink.runtime.messages.Acknowledge;
 import org.apache.flink.runtime.registration.RegistrationResponse;
 import org.apache.flink.runtime.resourcemanager.JobLeaderIdService;
-import org.apache.flink.runtime.resourcemanager.exceptions.ResourceManagerException;
 import org.apache.flink.runtime.resourcemanager.registration.JobManagerRegistration;
-import org.apache.flink.runtime.rpc.FatalErrorHandler;
-import org.apache.flink.runtime.rpc.FencedRpcEndpoint;
-import org.apache.flink.runtime.rpc.RpcService;
-import org.apache.flink.runtime.rpc.RpcUtils;
+import org.apache.flink.runtime.rpc.*;
 import org.apache.flink.runtime.rpc.akka.AkkaRpcServiceUtils;
 import org.apache.flink.runtime.webmonitor.retriever.LeaderGatewayRetriever;
 import org.apache.flink.streaming.controlplane.streammanager.exceptions.StreamManagerException;
@@ -177,7 +170,7 @@ public class StreamManager extends FencedRpcEndpoint<StreamManagerId> implements
 				jobLeaderIdService.addJob(jobId);
 			} catch (Exception e) {
 				StreamManagerException exception = new StreamManagerException("Could not add the job " +
-						jobId + " to the job id leader service.", e);
+					jobId + " to the job id leader service.", e);
 
 				onFatalError(exception);
 				log.error("Could not add job {} to job leader id service.", jobId, e);
@@ -195,7 +188,7 @@ public class StreamManager extends FencedRpcEndpoint<StreamManagerId> implements
 			// we cannot check the job leader id so let's fail
 			// TODO: Maybe it's also ok to skip this check in case that we cannot check the leader id
 			StreamManagerException exception = new StreamManagerException("Cannot obtain the " +
-					"job leader id future to verify the correct job leader.", e);
+				"job leader id future to verify the correct job leader.", e);
 
 			onFatalError(exception);
 
@@ -206,42 +199,42 @@ public class StreamManager extends FencedRpcEndpoint<StreamManagerId> implements
 		CompletableFuture<JobMasterGateway> jobMasterGatewayFuture = getRpcService().connect(jobManagerAddress, jobMasterId, JobMasterGateway.class);
 
 		CompletableFuture<RegistrationResponse> registrationResponseFuture = jobMasterGatewayFuture.thenCombineAsync(
-				jobMasterIdFuture,
-				(JobMasterGateway jobMasterGateway, JobMasterId leadingJobMasterId) -> {
-					if (Objects.equals(leadingJobMasterId, jobMasterId)) {
-						return registerJobMasterInternal(
-								jobMasterGateway,
-								jobId,
-								jobManagerAddress,
-								jobManagerResourceId);
-					} else {
-						final String declineMessage = String.format(
-								"The leading JobMaster id %s did not match the received JobMaster id %s. " +
-										"This indicates that a JobMaster leader change has happened.",
-								leadingJobMasterId,
-								jobMasterId);
-						log.debug(declineMessage);
-						return new RegistrationResponse.Decline(declineMessage);
-					}
-				},
-				getMainThreadExecutor());
+			jobMasterIdFuture,
+			(JobMasterGateway jobMasterGateway, JobMasterId leadingJobMasterId) -> {
+				if (Objects.equals(leadingJobMasterId, jobMasterId)) {
+					return registerJobMasterInternal(
+						jobMasterGateway,
+						jobId,
+						jobManagerAddress,
+						jobManagerResourceId);
+				} else {
+					final String declineMessage = String.format(
+						"The leading JobMaster id %s did not match the received JobMaster id %s. " +
+							"This indicates that a JobMaster leader change has happened.",
+						leadingJobMasterId,
+						jobMasterId);
+					log.debug(declineMessage);
+					return new RegistrationResponse.Decline(declineMessage);
+				}
+			},
+			getMainThreadExecutor());
 
 		// handle exceptions which might have occurred in one of the futures inputs of combine
 		return registrationResponseFuture.handleAsync(
-				(RegistrationResponse registrationResponse, Throwable throwable) -> {
-					if (throwable != null) {
-						if (log.isDebugEnabled()) {
-							log.debug("Registration of job manager {}@{} failed.", jobMasterId, jobManagerAddress, throwable);
-						} else {
-							log.info("Registration of job manager {}@{} failed.", jobMasterId, jobManagerAddress);
-						}
-
-						return new RegistrationResponse.Decline(throwable.getMessage());
+			(RegistrationResponse registrationResponse, Throwable throwable) -> {
+				if (throwable != null) {
+					if (log.isDebugEnabled()) {
+						log.debug("Registration of job manager {}@{} failed.", jobMasterId, jobManagerAddress, throwable);
 					} else {
-						return registrationResponse;
+						log.info("Registration of job manager {}@{} failed.", jobMasterId, jobManagerAddress);
 					}
-				},
-				getRpcService().getExecutor());
+
+					return new RegistrationResponse.Decline(throwable.getMessage());
+				} else {
+					return registrationResponse;
+				}
+			},
+			getRpcService().getExecutor());
 	}
 
 	@Override
@@ -258,17 +251,17 @@ public class StreamManager extends FencedRpcEndpoint<StreamManagerId> implements
 	/**
 	 * Registers a new JobMaster.
 	 *
-	 * @param jobMasterGateway to communicate with the registering JobMaster
-	 * @param jobId of the job for which the JobMaster is responsible
-	 * @param jobManagerAddress address of the JobMaster
+	 * @param jobMasterGateway     to communicate with the registering JobMaster
+	 * @param jobId                of the job for which the JobMaster is responsible
+	 * @param jobManagerAddress    address of the JobMaster
 	 * @param jobManagerResourceId ResourceID of the JobMaster
 	 * @return RegistrationResponse
 	 */
 	private RegistrationResponse registerJobMasterInternal(
-			final JobMasterGateway jobMasterGateway,
-			JobID jobId,
-			String jobManagerAddress,
-			ResourceID jobManagerResourceId) {
+		final JobMasterGateway jobMasterGateway,
+		JobID jobId,
+		String jobManagerAddress,
+		ResourceID jobManagerResourceId) {
 		if (jobManagerRegistration != null) {
 			JobManagerRegistration oldJobManagerRegistration = jobManagerRegistration;
 			if (Objects.equals(oldJobManagerRegistration.getJobMasterId(), jobMasterGateway.getFencingToken())) {
@@ -276,19 +269,19 @@ public class StreamManager extends FencedRpcEndpoint<StreamManagerId> implements
 				log.debug("Job manager {}@{} was already registered.", jobMasterGateway.getFencingToken(), jobManagerAddress);
 			} else {
 				disconnectJobManager(
-						oldJobManagerRegistration.getJobID(),
-						new Exception("New job leader for job " + jobId));
+					oldJobManagerRegistration.getJobID(),
+					new Exception("New job leader for job " + jobId));
 
 				this.jobManagerRegistration = new JobManagerRegistration(
-						jobId,
-						jobManagerResourceId,
-						jobMasterGateway);
-			}
-		} else {
-			this.jobManagerRegistration = new JobManagerRegistration(
 					jobId,
 					jobManagerResourceId,
 					jobMasterGateway);
+			}
+		} else {
+			this.jobManagerRegistration = new JobManagerRegistration(
+				jobId,
+				jobManagerResourceId,
+				jobMasterGateway);
 		}
 
 		log.info("Registered job manager {}@{} for job {}.", jobMasterGateway.getFencingToken(), jobManagerAddress, jobId);
@@ -296,8 +289,8 @@ public class StreamManager extends FencedRpcEndpoint<StreamManagerId> implements
 		// TODO: HeartBeatService
 
 		return new JobMasterRegistrationSuccess<StreamManagerId>(
-				getFencingToken(),
-				resourceId);
+			getFencingToken(),
+			resourceId);
 	}
 
 	protected void closeJobManagerConnection(JobID jobID, Exception cause) {
@@ -315,7 +308,7 @@ public class StreamManager extends FencedRpcEndpoint<StreamManagerId> implements
 				try {
 					log.info("connect successfully");
 					gateway.submitJob(jobGraph,
-						new StreamManagerAddress(this.getAddress(), getFencingToken()),
+						this.getAddress(),
 						Time.seconds(10));
 				} catch (Exception e) {
 					log.error("Error while invoking runtime dispatcher RMI.", e);
@@ -383,7 +376,8 @@ public class StreamManager extends FencedRpcEndpoint<StreamManagerId> implements
 	protected void onFatalError(Throwable t) {
 		try {
 			log.error("Fatal error occurred in ResourceManager.", t);
-		} catch (Throwable ignored) {}
+		} catch (Throwable ignored) {
+		}
 
 		// The fatal error handler implementation should make sure that this call is non-blocking
 		fatalErrorHandler.onFatalError(t);
