+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/controlplane/streammanager/StreamManagerService.java
@@ -16,8 +16,10 @@
  * limitations under the License.
  */
 
-package org.apache.flink.runtime.controlplane.streammanager;
+package org.apache.flink.streaming.controlplane.streammanager;
 
+import org.apache.flink.runtime.controlplane.streammanager.StreamManagerGateway;
+import org.apache.flink.runtime.controlplane.streammanager.StreamManagerId;
 import org.apache.flink.runtime.messages.Acknowledge;
 import org.apache.flink.util.AutoCloseableAsync;
