+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/controlplane/dispatcher/StreamManagerDispatcherTest.java
@@ -26,8 +26,7 @@ import org.apache.flink.runtime.blob.BlobServer;
 import org.apache.flink.runtime.blob.VoidBlobStore;
 import org.apache.flink.runtime.checkpoint.StandaloneCheckpointRecoveryFactory;
 import org.apache.flink.runtime.clusterframework.types.ResourceID;
-import org.apache.flink.runtime.controlplane.streammanager.StreamManagerRunner;
-import org.apache.flink.runtime.controlplane.streammanager.StreamManagerService;
+import org.apache.flink.streaming.controlplane.streammanager.StreamManagerRunner;
 import org.apache.flink.runtime.dispatcher.*;
 import org.apache.flink.runtime.heartbeat.HeartbeatServices;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServices;
@@ -55,9 +54,7 @@ import org.apache.flink.runtime.util.ExecutorThreadFactory;
 import org.apache.flink.runtime.util.Hardware;
 import org.apache.flink.runtime.util.TestingFatalErrorHandler;
 import org.apache.flink.runtime.webmonitor.retriever.LeaderGatewayRetriever;
-import org.apache.flink.runtime.webmonitor.retriever.impl.RpcGatewayRetriever;
 import org.apache.flink.streaming.controlplane.streammanager.*;
-import org.apache.flink.streaming.controlplane.streammanager.factories.DefaultStreamManagerServiceFactory;
 import org.apache.flink.streaming.controlplane.streammanager.factories.StreamManagerServiceFactory;
 import org.apache.flink.util.TestLogger;
 import org.apache.flink.util.function.ThrowingRunnable;
